requirements.txt
python-dotenv==1.0.0
discord.py==2.2.2
tomli==2.0.1
aiohttp==3.8.4